+++ dart/collision/native/narrow_phase/sphere_sphere.cpp
@@ -59,31 +59,44 @@ bool collideSpheresOnAxis(
 
   const double penetration = sumRadii - dist;
 
-  Eigen::Vector3d normal;
-  Eigen::Vector3d point;
   if (dist < 1e-10) {
-    normal = Eigen::Vector3d::UnitZ();
-    point = Eigen::Vector3d(center1X, center1Y, center1Z);
+    result.addContact(center1X, center1Y, center1Z, 0.0, 0.0, 1.0, penetration);
   } else {
     const double normalSign = (delta > 0.0) ? -1.0 : 1.0;
     const double pointOffset = normalSign * (-radius1 + penetration * 0.5);
     switch (axis) {
       case 0:
-        normal = Eigen::Vector3d(normalSign, 0.0, 0.0);
-        point = Eigen::Vector3d(center1X + pointOffset, center1Y, center1Z);
+        result.addContact(
+            center1X + pointOffset,
+            center1Y,
+            center1Z,
+            normalSign,
+            0.0,
+            0.0,
+            penetration);
         break;
       case 1:
-        normal = Eigen::Vector3d(0.0, normalSign, 0.0);
-        point = Eigen::Vector3d(center1X, center1Y + pointOffset, center1Z);
+        result.addContact(
+            center1X,
+            center1Y + pointOffset,
+            center1Z,
+            0.0,
+            normalSign,
+            0.0,
+            penetration);
         break;
       default:
-        normal = Eigen::Vector3d(0.0, 0.0, normalSign);
-        point = Eigen::Vector3d(center1X, center1Y, center1Z + pointOffset);
+        result.addContact(
+            center1X,
+            center1Y,
+            center1Z + pointOffset,
+            0.0,
+            0.0,
+            normalSign,
+            penetration);
         break;
     }
   }
-
-  result.addContact(point, normal, penetration);
   return true;
 }
 
@@ -103,15 +116,15 @@ bool collideSphereCenters(
   const double dy = center2.y() - center1.y();
   const double dz = center2.z() - center1.z();
 
-  if (dy == 0.0 && dz == 0.0 && std::isfinite(dx)) {
+  if (dy == 0.0 && dz == 0.0) {
     return collideSpheresOnAxis(
         center1.x(), center1.y(), center1.z(), radius1, radius2, dx, 0, result);
   }
-  if (dx == 0.0 && dz == 0.0 && std::isfinite(dy)) {
+  if (dx == 0.0 && dz == 0.0) {
     return collideSpheresOnAxis(
         center1.x(), center1.y(), center1.z(), radius1, radius2, dy, 1, result);
   }
-  if (dx == 0.0 && dy == 0.0 && std::isfinite(dz)) {
+  if (dx == 0.0 && dy == 0.0) {
     return collideSpheresOnAxis(
         center1.x(), center1.y(), center1.z(), radius1, radius2, dz, 2, result);
   }
@@ -177,7 +190,7 @@ bool collideSpheres(
   const double radius1 = detail::getRadius(sphere1);
   const double radius2 = detail::getRadius(sphere2);
 
-  if (dy == 0.0 && dz == 0.0 && std::isfinite(dx)) {
+  if (dy == 0.0 && dz == 0.0) {
     return collideSpheresOnAxis(
         translation1.x(),
         translation1.y(),
@@ -188,7 +201,7 @@ bool collideSpheres(
         0,
         result);
   }
-  if (dx == 0.0 && dz == 0.0 && std::isfinite(dy)) {
+  if (dx == 0.0 && dz == 0.0) {
     return collideSpheresOnAxis(
         translation1.x(),
         translation1.y(),
@@ -199,7 +212,7 @@ bool collideSpheres(
         1,
         result);
   }
-  if (dx == 0.0 && dy == 0.0 && std::isfinite(dz)) {
+  if (dx == 0.0 && dy == 0.0) {
     return collideSpheresOnAxis(
         translation1.x(),
         translation1.y(),